13/14 Category + Price display, delete offer dialog

+++ VivaCoronia/app/src/main/java/de/tudarmstadt/iptk/foxtrot/vivacoronia/trading/OfferOverviewFragment.kt
@@ -1,13 +1,16 @@
 package de.tudarmstadt.iptk.foxtrot.vivacoronia.trading
 
-import android.content.Intent
+import android.app.AlertDialog
+import android.graphics.Color
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.LinearLayout
 import android.widget.Toast
+import androidx.core.view.setMargins
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
@@ -55,20 +58,53 @@ class OffersFragment : Fragment() {
 
         GlobalScope.launch { fetchMyOffers() }
 
-        binding.add.setOnClickListener { AddOfferActivity.start(requireContext(), null) }
+        binding.add.setOnClickListener { SubmitOfferActivity.start(requireContext(), null) }
 
         return binding.root
     }
 
     private fun deleteOfferCallback(id: String) {
-        val deleted = TradingApiClient.deleteOffer(id)
-        if (deleted)
-            viewModel.remove(id)
-        // TODO better fetch again and reset whole list. Does not change efficiency because only differences matter
+        val dialogBuilder = activity?.let {
+            val builder = AlertDialog.Builder(it)
+            builder.apply {
+                setCancelable(true)
+                setPositiveButton(R.string.yes) {_, _ ->
+                    val deleted = TradingApiClient.deleteOffer(id)
+                    if (deleted)
+                        viewModel.remove(id)
+                    // TODO better fetch again and reset whole list. Does not change efficiency because only differences matter
+                }
+                setNegativeButton(R.string.cancel) { dialog, _ -> dialog.dismiss() }
+                setTitle(R.string.delete_offer_title)
+                setMessage(R.string.confirm_delete_message)
+            }
+        }
+            ?: return
+        val dialog = dialogBuilder.create()
+        dialog.show()
+        styleDeleteDialog(dialog) // TODO löschen wenn styling auch ohne so funktioniert wie bei Timo
+    }
+
+    private fun styleDeleteDialog(dialog: AlertDialog) {
+        val params = LinearLayout.LayoutParams(
+            LinearLayout.LayoutParams.WRAP_CONTENT,
+            LinearLayout.LayoutParams.WRAP_CONTENT
+        )
+        params.setMargins(5)
+
+        val buttons = listOf(
+            dialog.getButton(AlertDialog.BUTTON_NEGATIVE),
+            dialog.getButton(AlertDialog.BUTTON_POSITIVE)
+        )
+        for (button in buttons) {
+            button.setTextColor(Color.BLACK)
+            button.setBackgroundColor(Color.LTGRAY)
+            button.layoutParams = params
+        }
     }
 
     private fun editOfferCallback(offer: Offer){
-        AddOfferActivity.start(requireContext(), offer)
+        SubmitOfferActivity.start(requireContext(), offer)
     }
 
     private fun fetchMyOffers() {
